Balancing Cubes
There are N  cubes of sugar. All cubes of sugar are pure except one cube which is impure. The property states that the cube which is impure weighs less than that of pure. You have only one balance weighing machine that contains two sides so that you can compare the weight of things you put on both sides. In one operation, you can use a balanced weighing machine once. You are required to determine which cube is impure using a balanced weighing machine in minimum operation possible so that after all the operations you performed, you become sure that this piece of the cube is surely impure. 

Note: After that minimum operations, you have to be absolutely certain about that impure cube in the worst-case scenario.

Input format

The first and only line contains N denoting the number of cubes of sugar.